requirements.txt
acres==0.2.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
etelemetry==0.3.1
fastapi==0.115.12
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.6.1
jiter==0.8.2
looseversion==1.3.0
lxml==5.3.0
Mako==1.3.10
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.1
openai==1.84.0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pydantic==2.10.5
pydantic_core==2.27.2
pydot==3.0.4
PyMuPDF==1.26.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.15.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.3
